refactor: centralize timed checks
